+++ CHANGELOG.md
@@ -31,8 +31,7 @@
       (@arogachev)
 - Bug #54: Fix ignoring of using `Assignment::$createdAt` in `AssignmentsStorage::add()` (@arogachev)
 - Chg #57: Raise PHP version to 8.1 (@arogachev)
-- Chg #?: Add customizable separator for joining and splitting item names (@arogachev)
-- Chg #?: Raise PHP version to 8.1 (@arogachev)
+- Chg #59: Add customizable separator for joining and splitting item names (@arogachev)
 
 ## 1.0.0 April 20, 2023