+++ CHANGELOG.md
@@ -32,6 +32,7 @@
 - Bug #54: Fix ignoring of using `Assignment::$createdAt` in `AssignmentsStorage::add()` (@arogachev)
 - Chg #57: Raise PHP version to 8.1 (@arogachev)
 - Chg #59: Add customizable separator for joining and splitting item names (@arogachev)
+- Enh #60: Use migrations (@arogachev)
 
 ## 1.0.0 April 20, 2023